add: components分解
